refactor: 패키지와 메서드 리네이밍 및 불필요한 메서드 삭제
